requirements.txt
nbconvert
pyyaml
scholarly
requests
